pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-reviewer"
dynamic = ["version"]
description = "A Python project called git-reviewer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Ryan Nikolaidis", email = "ryannikolaidis@gmail.com" },
]
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "PyYAML>=6.0.0",
    "nllm @ git+https://github.com/ryannikolaidis/nllm.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "pyright>=1.1.300",
    "pre-commit>=3.6.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",]
[project.scripts]
git-reviewer = "git_reviewer.cli:main"
[project.urls]
Homepage = "https://github.com/ryannikolaidis/git-reviewer"
Repository = "https://github.com/ryannikolaidis/git-reviewer"
Issues = "https://github.com/ryannikolaidis/git-reviewer/issues"

[tool.hatch.version]
path = "git_reviewer/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["git_reviewer"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["git_reviewer"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
# Fast and accurate type checking
useLibraryCodeForTypes = true
# Exclude tests by default for faster checking
exclude = [
    "tests/",
    ".venv/",
    "build/",
    "dist/",
]
# Report settings
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=git_reviewer --cov-report=term-missing"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
